Product module updated
